feat(docs): need split examples of validation (#UIM-543) WIP
