+++ packages/mosaic-examples/mosaic/validation/validation-composite-overview/validation-composite-overview-example.html
@@ -1,0 +1,21 @@
+<div class="layout-margin dev-composite-form" [class.validation-error]="compositeForm.errors?.range">
+    <form class="dev-form" [formGroup]="compositeForm" novalidate>
+        <div class="layout-row">
+            <div class="dev-column">
+                <div class="dev-label">Начало диапазона</div>
+                <mc-form-field>
+                    <input mcInput formControlName="start">
+                </mc-form-field>
+            </div>
+
+            <div class="dev-column">
+                <div class="dev-label">Окончание</div>
+                <mc-form-field>
+                    <input mcInput formControlName="end">
+                </mc-form-field>
+            </div>
+        </div>
+
+        <mc-hint class="dev-hint" *ngIf="compositeForm.errors?.range">IP-адрес окончания диапазона должен быть больше, чем IP-адрес начала</mc-hint>
+    </form>
+</div>
