+++ packages/mosaic-examples/mosaic/validation/validation-global-overview/validation-global-overview-example.html
@@ -1,0 +1,34 @@
+<div class="layout-margin dev-width">
+    <div class="mc-alert mc-alert_error" *ngIf="showServerErrors">
+        <i mc-icon="mc-info_16" color="error" class="mc-alert__icon"></i>
+        <div>
+            <div><b>Не удается получить ответ от сервера</b></div>
+            При попытке отправить данные сервер не ответил, попробуйте снова
+        </div>
+    </div>
+
+    <form class="dev-form" [formGroup]="globalErrorForm" (ngSubmit)="submitGlobalErrorForm()" novalidate>
+        <div class="dev-row">
+            <div class="dev-label">Имя</div>
+            <mc-form-field>
+                <input mcInput formControlName="firstName">
+            </mc-form-field>
+        </div>
+
+        <div class="dev-row">
+            <div class="dev-label">Фамилия</div>
+            <mc-form-field>
+                <input mcInput formControlName="lastName">
+            </mc-form-field>
+        </div>
+
+        <button
+            mc-button
+            color="primary"
+            type="submit"
+            [class.mc-progress]="inProgress"
+            [disabled]="inProgress">
+            Отправить
+        </button>
+    </form>
+</div>
