+++ packages/mosaic-examples/mosaic/validation/validation-overview/validation-overview-example.html
@@ -1,5 +1,3 @@
-
-Валидация обязательных полей
 <div class="layout-row layout-margin">
     <div class="dev-width">
         С текстом ошибок
@@ -134,121 +132,4 @@
             <button mc-button color="primary" type="submit">Отправить</button>
         </form>
     </div>
-</div>
-
-По уходу с поля
-<div class="layout-margin dev-width">
-    <div class="dev-row">
-        <div class="dev-label">IP-адрес</div>
-        <mc-form-field>
-            <input mcInput [formControl]="ipAddress">
-            <mc-hint class="dev-hint" *ngIf="ipAddress.invalid">IP-адрес не соответствует стандартам RFC</mc-hint>
-        </mc-form-field>
-    </div>
-</div>
-
-<br><br><br><br><br>
-
-Глобальная ошибка
-<div class="layout-margin dev-width">
-    <div class="mc-alert mc-alert_error" *ngIf="showServerErrors">
-        <i mc-icon="mc-info_16" color="error" class="mc-alert__icon"></i>
-        <div>
-            <div><b>Не удается получить ответ от сервера</b></div>
-            При попытке отправить данные сервер не ответил, попробуйте снова
-        </div>
-    </div>
-
-    <form class="dev-form" [formGroup]="globalErrorForm" (ngSubmit)="submitGlobalErrorForm()" novalidate>
-        <div class="dev-row">
-            <div class="dev-label">Имя</div>
-            <mc-form-field>
-                <input mcInput formControlName="firstName">
-            </mc-form-field>
-        </div>
-
-        <div class="dev-row">
-            <div class="dev-label">Фамилия</div>
-            <mc-form-field>
-                <input mcInput formControlName="lastName">
-            </mc-form-field>
-        </div>
-
-        <button
-            mc-button
-            color="primary"
-            type="submit"
-            [class.mc-progress]="inProgress"
-            [disabled]="inProgress">
-            Отправить
-        </button>
-    </form>
-</div>
-
-<br><br><br><br><br>
-
-Обязательные поля на небольших формах
-<div class="layout-margin dev-width">
-    <form class="dev-form" [formGroup]="smallForm" novalidate>
-        <div class="dev-row">
-            <div class="dev-label">Имя</div>
-            <mc-form-field>
-                <input mcInput formControlName="firstName">
-            </mc-form-field>
-        </div>
-
-        <div class="dev-row">
-            <div class="dev-label">Фамилия</div>
-            <mc-form-field>
-                <input mcInput formControlName="lastName">
-            </mc-form-field>
-        </div>
-
-        <button mc-button color="primary" type="submit" [disabled]="smallForm.errors?.empty">Отправить</button>
-    </form>
-</div>
-
-<br><br><br><br><br>
-
-Во время ввода значения
-<div class="layout-margin dev-width">
-    <form class="dev-form" [formGroup]="checkOnFlyForm" novalidate>
-        <div class="dev-row">
-            <div class="dev-label">Название папки</div>
-            <mc-form-field
-                mcTooltip
-                [mcTitle]="'Только цифры'"
-                #tooltip="mcTooltip"
-                mcTrigger="manual"
-                mcPlacement="bottom">
-                <input mcInput formControlName="folderName" (input)="onInput($event)">
-            </mc-form-field>
-        </div>
-    </form>
-</div>
-
-<br><br><br><br><br>
-
-Составные поля
-
-<div class="layout-margin dev-composite-form" [class.validation-error]="compositeForm.errors?.range">
-    <form class="dev-form" [formGroup]="compositeForm" novalidate>
-        <div class="layout-row">
-            <div class="dev-column">
-                <div class="dev-label">Начало диапазона</div>
-                <mc-form-field>
-                    <input mcInput formControlName="start">
-                </mc-form-field>
-            </div>
-
-            <div class="dev-column">
-                <div class="dev-label">Окончание</div>
-                <mc-form-field>
-                    <input mcInput formControlName="end">
-                </mc-form-field>
-            </div>
-        </div>
-
-        <mc-hint class="dev-hint" *ngIf="compositeForm.errors?.range">IP-адрес окончания диапазона должен быть больше, чем IP-адрес начала</mc-hint>
-    </form>
 </div>
