+++ packages/mosaic-examples/mosaic/validation/validation-small-overview/validation-small-overview-example.html
@@ -1,0 +1,19 @@
+<div class="layout-margin dev-width">
+    <form class="dev-form" [formGroup]="smallForm" novalidate>
+        <div class="dev-row">
+            <div class="dev-label">Имя</div>
+            <mc-form-field>
+                <input mcInput formControlName="firstName">
+            </mc-form-field>
+        </div>
+
+        <div class="dev-row">
+            <div class="dev-label">Фамилия</div>
+            <mc-form-field>
+                <input mcInput formControlName="lastName">
+            </mc-form-field>
+        </div>
+
+        <button mc-button color="primary" type="submit" [disabled]="smallForm.errors?.empty">Отправить</button>
+    </form>
+</div>
